Fix logout issue
