+++ src/components/Dropdown/index.js
@@ -31,7 +31,7 @@ const Dropdown = ({ options }) => {
 
         {showOptions && (
           <div
-            className="origin-top-right absolute right-0 mt-2 w-32 rounded-md shadow-lg bg-primary ring-1 ring-black ring-opacity-5 focus:outline-none"
+            className="origin-top-right absolute right-0 mt-4 w-32 rounded-md shadow-lg bg-primary ring-1 ring-black ring-opacity-5 focus:outline-none"
             role="menu"
             aria-orientation="vertical"
             aria-labelledby="menubutton"
