+++ src/components/Footer/index.js
@@ -7,7 +7,7 @@ const Footer = () => {
       <div className="fixed bottom-0 bg-primary w-full flex flex-row items-center justify-start gap-20 h-20 md:px-48 lg:px-64">
         <div className="flex flex-row items-center justify-between gap-5 text-main">
           <p className="text-xs">Made with ♥ by Alina!</p>
-          <a href="https://github.com/alshei" target="_blank">
+          <a href="https://github.com/mlkywy" target="_blank" rel="noreferrer">
             <div>
               <GithubLogo size="1.5rem" />
             </div>
@@ -16,13 +16,14 @@ const Footer = () => {
           <a
             href="https://www.linkedin.com/in/alina-sheikh-a59a4518a/"
             target="_blank"
+            rel="noreferrer"
           >
             <div>
               <LinkedinLogo size="1.5rem" />
             </div>
           </a>
 
-          <a href="https://twitter.com/wymlky" target="_blank">
+          <a href="https://twitter.com/wymlky" target="_blank" rel="noreferrer">
             <div>
               <TwitterLogo size="1.5rem" />
             </div>
@@ -31,7 +32,7 @@ const Footer = () => {
 
         <div className="flex flex-row items-center justify-between gap-5 text-accent">
           <a
-            href="https://github.com/alshei/notepad-v2"
+            href="https://github.com/mlkywy/spotify"
             target="_blank"
             rel="noopener noreferrer"
           >
