+++ src/components/Header/index.js
@@ -1,8 +1,21 @@
 import React from "react";
-import logo from "../../assets/logo.svg";
-import Dropdown from "../Dropdown";
-import { NavLink } from "../Buttons";
 import { Link } from "react-router-dom";
+import logo from "../../assets/logo.svg";
+
+// Components
+import Dropdown from "../Dropdown";
+import { Primary, NavLink } from "../Buttons";
+
+// Constants
+import {
+  SPOTIFY_ENDPOINTS,
+  REDIRECT_URI,
+  RESPONSE_TYPE,
+  SPOTIFY_SCOPES,
+} from "../../api/constants";
+
+// Hooks
+import useUser from "../../hooks/useUser";
 
 const options = [
   { label: "dark", value: "dark" },
@@ -11,6 +24,13 @@ const options = [
 ];
 
 const Header = () => {
+  const { token, logout } = useUser();
+
+  const scopes = [
+    SPOTIFY_SCOPES.PLAYLIST_MODIFY_PRIVATE,
+    SPOTIFY_SCOPES.PLAYLIST_MODIFY_PUBLIC,
+  ];
+
   return (
     <div className="fixed w-full flex flex-row items-center justify-between text-main h-20 z-10 md:px-32 lg:px-64">
       <div>
@@ -20,6 +40,18 @@ const Header = () => {
       </div>
 
       <div className="flex flex-row items-center gap-5 justify-between">
+        {!token ? (
+          <Primary
+            option="login to spotify"
+            link={`${SPOTIFY_ENDPOINTS.AUTHORIZE}?client_id=${
+              process.env.REACT_APP_CLIENT_ID
+            }&redirect_uri=${REDIRECT_URI}&scope=${scopes.join(
+              "%20"
+            )}&response_type=${RESPONSE_TYPE}&show_dialog=true`}
+          />
+        ) : (
+          <Primary option="logout of spotify" onClick={logout} />
+        )}
         <NavLink option="documentation" link="documentation" />
         <div>
           <Dropdown options={options} />
